requirements.txt
torch
transformers
streamlit
replicate
scikit-learn
langchain
pypdf
sentence_transformers
chromadb
pyrebase4
pymysql
joblib
jieba
scipy
googletrans
nltk
firebase-admin
plotly